Moved all recommendation-related code into a single file.
The various similary classes have been put inside the recommendation.rb
file. That file now declares a module named 'Recommendation' and the
class previously named 'Recommendation' has been re-named to Match.

I think if it's possible, we should try to avoid file-explosion and
limit the code for a particular chapter/concept to a single file.